Fix a typing issue that made the presence system not work

+++ inc/physics.h
@@ -103,7 +103,7 @@ void PHY_update();
 void PHY_end();
 void PHY_clean_presence();
 void PHY_set_presence(u8 ind);
-u8 PHY_is_occupied(u8 ind);
+u16 PHY_is_occupied(u8 ind);
 
 void kill(Actor * act, u8 speed_x, u8 speed_y);
